generateEntity(#37) : Modification de la classe pour générer une entité et ajout des repository Admin et Reader
