--- Framework/HydratorTrait.php
@@ -1,21 +1,0 @@
-<?php
-namespace Framework;
-
-trait HydratorTrait
-{    
-    /**
-     * hydrate
-     *
-     * @param  array $data
-     */
-    public function hydrate(array $data)
-    {
-        foreach ($data as $key => $value) {
-            $method = 'set'.ucfirst($key);
-
-            if (is_callable([$this, $method])) {
-                $this->$method($value);
-            }
-        }
-    }
-}
