+++ Framework/Manager/EntityManager.php
@@ -2,9 +2,21 @@
 
 namespace Framework\Manager;
 
-use Framework\HydratorTrait;
-
 class EntityManager
 {
-    use HydratorTrait;
+    /**
+     * hydrate
+     *
+     * @param  array $data
+     */
+    public function hydrate(array $data)
+    {
+        foreach ($data as $key => $value) {
+            $method = 'set'.ucfirst($key);
+
+            if (is_callable([$this, $method])) {
+                $this->$method($value);
+            }
+        }
+    }
 }
